requirements.txt
streamlit
asgiref==3.8.1
contourpy==1.2.1
cycler==0.12.1
Django==5.0.7
djangorestframework==3.15.2
fonttools==4.53.1
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
tzdata==2024.1
wheel==0.43.0